% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucbthesis.R
\name{ucbthesis}
\alias{ucbthesis}
\title{UC Berkeley thesis format.}
\usage{
ucbthesis(fig_width = 7, fig_height = 5, highlight = "pygments",
  citation_package = c("biblatex", "natbib"), ...)
}
\arguments{
\item{fig_width}{Default width (in inches) for figures.}

\item{fig_height}{Default width (in inches) for figures.}

\item{highlight}{Syntax highlighting style.}

\item{citation_package}{The LaTeX package to process citations, \code{natbib} or \code{biblatex}. The default in the R Markdown ucbthesis format is \code{biblatex}.}

\item{...}{Other arguments to \code{rmarkdown::pdf_document}.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating a University of California, Berkeley thesis or dissertation PDF document.
Adapted from
\href{https://math.berkeley.edu/~vojta/tex/ucbthesis-phd.html}{https://math.berkeley.edu/~vojta/tex/ucbthesis-phd.html} and
\href{https://math.berkeley.edu/~vojta/tex/ucbthesis-masters.html}{https://math.berkeley.edu/~vojta/tex/ucbthesis-masters.html}.
See the Graduate Division's formatting rules for theses (http://grad.berkeley.edu/academic-progress/thesis/)
and dissertations (http://grad.berkeley.edu/academic-progress/dissertation/) for more information.
}
\examples{

\dontrun{
rmarkdown::draft("my_thesis.Rmd", template = "ucbthesis", package = "ucbthesisrmd")
}

}

